+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+## [2.8] - 2020-07-08
 ### Added
 - `label_map` attribute
 - `positions` attribute
@@ -496,7 +497,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - TPR
 - documents and `README.md`
 
-[Unreleased]: https://github.com/sepandhaghighi/pycm/compare/v2.7...dev
+[Unreleased]: https://github.com/sepandhaghighi/pycm/compare/v2.8...dev
+[2.8]: https://github.com/sepandhaghighi/pycm/compare/v2.7...v2.8
 [2.7]: https://github.com/sepandhaghighi/pycm/compare/v2.6...v2.7
 [2.6]: https://github.com/sepandhaghighi/pycm/compare/v2.5...v2.6
 [2.5]: https://github.com/sepandhaghighi/pycm/compare/v2.4...v2.5